# NodeSetEditor.Server/OpenApi/v1.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "OPC UA NodeSet Editor API",
    "description": "REST API for the OPC UA NodeSet Editor.\n\nMost endpoints act on a workspace identified by the `OpcUa-Server` header, and NodeIds appearing in a URL path are base64url-encoded rather than percent-encoded.\n\nThis document is generated from the controllers at build time and checked into the repository; it is not produced at runtime.",
    "contact": {
      "name": "OPC Foundation",
      "url": "https://opcfoundation.org"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/opcua/v1/servers/{applicationUri}": {
      "get": {
        "tags": [
          "Editor"
        ],
        "parameters": [
          {
            "name": "applicationUri",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceDescription"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceDescription"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceDescription"
                }
              }
            }
          }
        },
        "security": [
          {
            "AzureAd": [ ]
          },
          {
            "EmailCookie": [ ]
          }
        ]
      },
      "put": {
        "tags": [
          "Editor"
        ],
        "parameters": [
          {
            "name": "applicationUri",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateServerRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateServerRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateServerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceDescription"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceDescription"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceDescription"
                }
              }
            }
          }
        },
        "security": [
          {
            "AzureAd": [ ]
          },
          {
            "EmailCookie": [ ]
          }
        ]
      },
      "delete": {
        "tags": [
          "Editor"
        ],
        "parameters": [
          {
            "name": "applicationUri",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "AzureAd": [ ]
          },
          {
            "EmailCookie": [ ]
          }
        ]
      }
    },
    "/api/opcua/v1/servers": {
      "post": {
        "tags": [
          "Editor"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateServerRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateServerRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateServerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceDescription"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceDescription"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceDescription"
                }
              }
            }
          }
        },
        "security": [
          {
            "AzureAd": [ ]
          },
          {
            "EmailCookie": [ ]
          }
        ]
      }
    },
...
    "/api/opcua/v1/nodes/{nodeId}": {
      "get": {
        "tags": [
          "Editor"
        ],
        "parameters": [
          {
            "name": "nodeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "OpcUa-Server",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Node"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Node"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Node"
                }
              }
            }
          }
        },
        "security": [
          {
            "AzureAd": [ ]
          },
          {
            "EmailCookie": [ ]
          }
        ]
      },
      "put": {
        "tags": [
          "Editor"
        ],
        "parameters": [
          {
            "name": "nodeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "OpcUa-Server",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNodeRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNodeRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Node"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Node"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Node"
                }
              }
            }
          }
        },
        "security": [
          {
            "AzureAd": [ ]
          },
          {
            "EmailCookie": [ ]
          }
        ]
      },
      "delete": {
        "tags": [
          "Editor"
        ],
        "parameters": [
          {
            "name": "nodeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "OpcUa-Server",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "AzureAd": [ ]
          },
          {
            "EmailCookie": [ ]
          }
        ]
      }
    },
...
    "/api/opcua/v1/nodes": {
      "post": {
        "tags": [
          "Editor"
        ],
        "parameters": [
          {
            "name": "OpcUa-Server",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNodeRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNodeRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Node"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Node"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Node"
                }
              }
            }
          }
        },
        "security": [
          {
            "AzureAd": [ ]
          },
          {
            "EmailCookie": [ ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateServerRequest": {
        "type": "object",
        "properties": {
          "applicationName": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "acl": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        }
      },
...
      "LocalizedText": {
        "type": "object",
        "properties": {
          "locale": {
            "type": "string",
            "nullable": true
          },
          "text": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "ServerConfiguration": {
        "type": "object",
        "properties": {
          "endpointUrl": {
            "type": "string",
            "nullable": true
          },
          "transportProfileUri": {
            "type": "string",
            "nullable": true
          },
          "securityMode": {
            "type": "string",
            "nullable": true
          },
          "securityPolicyUri": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "UpdateServerRequest": {
        "type": "object",
        "properties": {
          "applicationName": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "acl": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        }
      },
...
      "WorkspaceDescription": {
        "type": "object",
        "properties": {
          "description": {
            "oneOf": [
              {
                "enum": [
                  null
                ]
              },
              {
                "$ref": "#/components/schemas/LocalizedText"
              }
            ]
          },
          "serverConfiguration": {
            "oneOf": [
              {
                "enum": [
                  null
                ]
              },
              {
                "$ref": "#/components/schemas/ServerConfiguration"
              }
            ]
          },
          "owner": {
            "type": "string",
            "nullable": true
          },
          "isOwner": {
            "type": "boolean",
            "nullable": true
          },
          "canWrite": {
            "type": "boolean",
            "nullable": true
          },
          "acl": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "modifiedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "applicationUri": {
            "type": "string",
            "nullable": true
          },
          "productUri": {
            "type": "string",
            "nullable": true
          },
          "applicationName": {
            "oneOf": [
              {
                "enum": [
                  null
                ]
              },
              {
                "$ref": "#/components/schemas/LocalizedText"
              }
            ]
          },
          "discoveryUrls": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "isDefault": {
            "type": "boolean",
            "nullable": true
          }
        }
      },
...
    },
    "securitySchemes": {
      "AzureAd": {
        "type": "http",
        "description": "Azure AD access token — the \"sign in with Microsoft\" path.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "EmailCookie": {
        "type": "apiKey",
        "description": "Session cookie issued by the passwordless email-code sign-in (`POST /api/auth/verify-code`). Sent automatically by the browser.",
        "name": "opc-email-auth",
        "in": "cookie"
      }
    }
  },
...
}